# Fix home N+1s for users/speakers