@adrienpoly 2025-08-29
thanks, I like to get rid of those n+1